Deprecated dynamic resolver module to use cas ext
Deprecate `cas-security-dynamic-service-resolver`, now please use `spring-security-cas-extension`

I did that change since I have to extend _Spring security cas_ not only to support dynamic resolver. So I needed a more generic module name.

closes #34
