+++ cas-security-spring-boot-autoconfigure/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.kakawait</groupId>
     <artifactId>cas-security-spring-boot-autoconfigure</artifactId>
-    <version>0.3.1</version>
+    <version>0.4.0</version>
     <packaging>jar</packaging>
 
     <name>Cas security spring boot autoconfigure</name>
@@ -44,8 +44,8 @@
 
         <spring-boot-dependencies.version>1.5.4.RELEASE</spring-boot-dependencies.version>
 
-        <cas-security-dynamic-service-resolver.version>0.3.0</cas-security-dynamic-service-resolver.version>
-        <lombok.version>1.16.16</lombok.version>
+        <spring-security-cas-extension.version>0.1.0</spring-security-cas-extension.version>
+        <lombok.version>1.16.18</lombok.version>
         <hibernate-validator.version>5.4.1.Final</hibernate-validator.version>
 
         <maven-source-plugin.version>3.0.1</maven-source-plugin.version>
@@ -72,24 +72,15 @@
             <artifactId>spring-boot-autoconfigure</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.security</groupId>
-            <artifactId>spring-security-cas</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.kakawait</groupId>
-            <artifactId>cas-security-dynamic-service-resolver</artifactId>
-            <version>${cas-security-dynamic-service-resolver.version}</version>
+            <artifactId>spring-security-cas-extension</artifactId>
+            <version>${spring-security-cas-extension.version}</version>
         </dependency>
         <dependency>
             <groupId>org.hibernate</groupId>
             <artifactId>hibernate-validator</artifactId>
             <version>${hibernate-validator.version}</version>
         </dependency>
-
 
         <dependency>
             <groupId>org.projectlombok</groupId>
