+++ pom.xml
@@ -5,11 +5,11 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.kakawait</groupId>
-    <artifactId>cas-security-spring-boot-starter-parent</artifactId>
+    <artifactId>cas-security-spring-boot-parent</artifactId>
     <packaging>pom</packaging>
-    <version>0.3.1</version>
+    <version>0.4.0</version>
 
-    <name>Cas security spring boot starter parent</name>
+    <name>Cas security spring boot parent</name>
     <description>Spring boot starter for Apereo CAS client fully integrated with Spring security</description>
     <url>https://github.com/kakawait/cas-security-spring-boot-starter</url>
 
@@ -39,9 +39,9 @@
     </properties>
 
     <modules>
+        <module>spring-security-cas-extension</module>
         <module>cas-security-spring-boot-starter</module>
         <module>cas-security-spring-boot-autoconfigure</module>
-        <module>cas-security-dynamic-service-resolver</module>
     </modules>
 
 </project>
