+++ spring-security-cas-extension/src/main/java/com/kakawait/spring/security/cas/LaxServiceProperties.java
@@ -1,13 +1,11 @@
-package com.kakawait.security.cas;
+package com.kakawait.spring.security.cas;
 
 import org.springframework.security.cas.ServiceProperties;
 import org.springframework.util.Assert;
 
 /**
  * @author Thibaud Leprêtre
- * @deprecated Please use new artifact id {@code spring-security-cas-extension}
  */
-@Deprecated
 public class LaxServiceProperties extends ServiceProperties {
 
     @Override
